requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-multipart
boto3
botocore
pydantic>=2.8.0
python-dotenv
httpx
celery[redis]
redis
pytest
pytest-asyncio
aiofiles
python-docx
pdfminer.six
rapidfuzz
passlib[bcrypt]
redis[asyncio]
motor==3.1.1
beanie==2.0.1
pydantic-settings>=0.2.1
dnspython==2.4.2

# NumPy compatibility for LangChain (Python 3.12+)
numpy>=1.26.0,<2.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.24.1
python-multipart==0.0.6    # if file uploads not already present
